+++ frontend/src/App.tsx
@@ -1,9 +1,8 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
-import AdminMoviesPage from './completedPagesJD/AdminMoviesPage';
-import Homepage from './HomepageComplete/Homepage';
-import { useState } from 'react';
-      
+import AdminMoviesPage from './pages/AdminMoviesPage';
+import Homepage from './pages/Homepage';
+
 function App() {
   return (
     <>
@@ -11,6 +10,7 @@ function App() {
         <Routes>
           <Route path='adminmovies' element={<AdminMoviesPage />} />
           <Route path='/home' element={<Homepage />} />
+          <Route path='/' element={<Homepage />} />
         </Routes>
       </Router>
     </>
