+++ frontend/src/components/FilmJourneyCarousel.tsx
