+++ frontend/src/components/HeroBanner.tsx
