+++ frontend/src/components/MoodBoardGrid.tsx
