+++ frontend/src/components/MovieFilter.tsx
