+++ frontend/src/components/NewMovieForm.tsx
